+++ src/main/java/cart/auth/AuthInterceptor.java
@@ -1,6 +1,5 @@
 package cart.auth;
 
-import cart.exeception.AuthorizationException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpHeaders;
