+++ src/main/java/cart/auth/AuthorizationException.java
@@ -1,4 +1,4 @@
-package cart.exeception;
+package cart.auth;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
