+++ src/main/java/cart/auth/BasicCredentialExtractor.java
@@ -1,6 +1,5 @@
 package cart.auth;
 
-import cart.exeception.AuthorizationException;
 import org.apache.tomcat.util.codec.binary.Base64;
 import org.springframework.stereotype.Component;
